+++ src/lib/djerba/plugins/base.py
@@ -17,12 +17,7 @@ class plugin_base(configurable, ABC):
         super().__init__(**kwargs)
         self.workspace = kwargs['workspace']
         self.json_validator = plugin_json_validator(self.log_level, self.log_path)
-        defaults = {
-            core_constants.CONFIGURE_PRIORITY: self.DEFAULT_CONFIG_PRIORITY,
-            core_constants.EXTRACT_PRIORITY: self.DEFAULT_CONFIG_PRIORITY,
-            core_constants.RENDER_PRIORITY: self.DEFAULT_CONFIG_PRIORITY
-        }
-        self.set_all_ini_defaults(defaults)
+        self.specify_params()
 
     # configure() method is defined in parent class
 
@@ -51,3 +46,7 @@ class plugin_base(configurable, ABC):
         self.logger.debug(msg)
         self.json_validator.validate_data(data)
         return ''
+
+    def set_priority_defaults(self, priority):
+        for key in core_constants.PRIORITY_KEYS:
+            self.ini_defaults[key] = priority